front/requirements.txt
uvicorn==0.21.1
pydantic==1.10.7
fastapi==0.95.1
python-dotenv==1.0.0
python-json-logger==2.0.7
requests==2.28.2
protobuf==4.22.3
opentelemetry-api==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-instrumentation-botocore==0.38b0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-instrumentation-requests==0.38b0
opentelemetry-instrumentation-logging==0.38b0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-sdk-extension-aws==2.0.1
opentelemetry-propagator-aws-xray==1.0.1